clean up a little
