+++ src/pyhs3/allexjsontest.py
@@ -1,14 +1,16 @@
 from __future__ import annotations
 
 import json
+import logging
 import math
 from collections import OrderedDict
 
 import networkx as nx
 import numpy as np
 import pytensor.tensor as pt
-from pytensor import function as function
 from pytensor.graph.basic import graph_inputs
+
+log = logging.getLogger(__name__)
 
 json_content = r"""
 {
@@ -232,7 +234,6 @@ class Workspace:
             Model: The constructed model object.
         """
 
-        distlist = self.distribution_set
         assert set(parameter_point.points.keys()) == set(domain.domains.keys()), (
             "parameter and domain names do not match"
         )
@@ -303,7 +304,7 @@ class Model:
         Returns:
             float: The evaluated PDF value.
         """
-        print(parametervalues)
+        log.info(parametervalues)
 
         dist = self.distributions[name]
         return dist.eval(
@@ -418,7 +419,7 @@ class DomainCollection:
         domains (OrderedDict): Mapping of domain names to DomainSet objects.
     """
 
-    def __init__(self, domainsets: [DomainSet]):
+    def __init__(self, domainsets: list[DomainSet]):
         self.domains = OrderedDict()
 
         for domain_config in domainsets:
@@ -542,7 +543,7 @@ class GaussianDist(Distribution):
         Returns:
             pt.TensorVariable: Symbolic representation of the Gaussian PDF.
         """
-        # print("parameters: ", parameters)
+        # log.info("parameters: ", parameters)
         norm_const = 1.0 / (
             pt.sqrt(2 * math.pi) * distributionsandparameters[self.sigma]
         )
@@ -667,20 +668,7 @@ mymodel = myworkspace.model(
 )
 
 
-scalarranges = myworkspace.domain_collection["default_domain"]
-
-f = boundedscalar("f", scalarranges["f"])
-f_ctl = boundedscalar("f_ctl", scalarranges["f_ctl"])
-mean = boundedscalar("mean", scalarranges["mean"])
-mean2 = boundedscalar("mean2", scalarranges["mean2"])
-sigma = boundedscalar("sigma", scalarranges["sigma"])
-sigma2 = boundedscalar("sigma2", scalarranges["sigma2"])
-mean_ctl = boundedscalar("mean_ctl", scalarranges["mean_ctl"])
-mean2_ctl = boundedscalar("mean2_ctl", scalarranges["mean2_ctl"])
-# breakpoint()
-x = boundedscalar("x", scalarranges["x"])
-
-print("f: ", mymodel.parameterset["f"].value)
+log.info("f: %f", mymodel.parameterset["f"].value)
 
 physicspdfval = mymodel.pdf(
     "model",
@@ -700,95 +688,5 @@ physicspdfvalctl = mymodel.pdf(
     sigma=mymodel.parameterset["sigma"].value,
 )
 
-print(physicspdfval)
-print(physicspdfvalctl)
-
-
-def gaussian_pdf(x, mu, sigma):
-    norm_const = 1.0 / (pt.sqrt(2 * math.pi) * sigma)
-    exponent = pt.exp(-0.5 * ((x - mu) / sigma) ** 2)
-    return norm_const * exponent
-
-
-def mixture_pdf(coeff, pdf1, pdf2):
-    return coeff * pdf1 + (1.0 - coeff) * pdf2
-
-
-print("distributions: ", mymodel.distributions)
-gx = gaussian_pdf(x, mean, sigma)
-px = gaussian_pdf(x, mean2, sigma2)
-model = mixture_pdf(f, gx, px)
-
-gx_ctl = gaussian_pdf(x, mean_ctl, sigma)
-px_ctl = gaussian_pdf(x, mean2_ctl, sigma)
-model_ctl = mixture_pdf(f_ctl, gx_ctl, px_ctl)
-
-sample = pt.scalar("sample", dtype="int32")
-
-simPdf = pt.switch(pt.eq(sample, 0), model, model_ctl)
-
-pdf_physics = function([x, f, mean, sigma, mean2, sigma2], model, name="pdf_physics")
-
-pdf_control = function(
-    [x, f_ctl, mean_ctl, mean2_ctl, sigma], model_ctl, name="pdf_control"
-)
-
-pdf_combined = function(
-    [sample, x, f, mean, sigma, mean2, sigma2, f_ctl, mean_ctl, mean2_ctl],
-    simPdf,
-    name="pdf_combined",
-)
-
-# default_params = {p["name"]: p["value"] for p in mymodel.startingpoints}
-# default_params = mymodel.parameterset
-# mymodel.parameters['x']
-# breakpoint()
-val_physics = pdf_physics(
-    mymodel.parameterset["x"].value,
-    mymodel.parameterset["f"].value,
-    mymodel.parameterset["mean"].value,
-    mymodel.parameterset["sigma"].value,
-    mymodel.parameterset["mean2"].value,
-    mymodel.parameterset["sigma2"].value,
-)
-
-val_control = pdf_control(
-    mymodel.parameterset["x"].value,
-    mymodel.parameterset["f_ctl"].value,
-    mymodel.parameterset["mean_ctl"].value,
-    mymodel.parameterset["mean2_ctl"].value,
-    mymodel.parameterset["sigma"].value,
-)
-
-val_combined_physics = pdf_combined(
-    0,  # sample
-    mymodel.parameterset["x"].value,
-    mymodel.parameterset["f"].value,
-    mymodel.parameterset["mean"].value,
-    mymodel.parameterset["sigma"].value,
-    mymodel.parameterset["mean2"].value,
-    mymodel.parameterset["sigma2"].value,
-    mymodel.parameterset["f_ctl"].value,
-    mymodel.parameterset["mean_ctl"].value,
-    mymodel.parameterset["mean2_ctl"].value,
-)
-
-val_combined_control = pdf_combined(
-    1,  # sample
-    mymodel.parameterset["x"].value,
-    mymodel.parameterset["f"].value,
-    mymodel.parameterset["mean"].value,
-    mymodel.parameterset["sigma"].value,
-    mymodel.parameterset["mean2"].value,
-    mymodel.parameterset["sigma2"].value,
-    mymodel.parameterset["f_ctl"].value,
-    mymodel.parameterset["mean_ctl"].value,
-    mymodel.parameterset["mean2_ctl"].value,
-)
-
-print("Physics PDF:", val_physics)
-print("Control PDF:", val_control)
-print("Simultaneous PDF:", val_combined_physics)
-print("Simultaneous PDF(control):", val_combined_control)
-
-print(mymodel.parameterset)
+log.info(physicspdfval)
+log.info(physicspdfvalctl)
